3.x.8 (#228)
* unified wait type exclusion for the xes collector

* log output tweak

* added errorhandler to Task.WhenAll

* updated XES. DROP THE EXISTING SESSION BEFORE UPGRADING

DROP EVENT SESSION [SQLWATCH_waits] ON SERVER
GO

* date format in the buildversion info

* fixed file size

* redesigned XES wait stats for better perfromance, less storage utilisation and for better integration with grafana

* project maint.

* tweaked messaging

* added example keys

* function to retrieve @@SERVERNAME in a common way to avoid implicit conversion.

* performance improvements

* performance improvement to avoid implicit conversion

* new procedure stats collector

* comments

* version bump to let SqlWatchImport know about the changes

* tidy up leftover variables from performance tuning

* test

* test

* solved the merge mismatch problem caused by VS encoding

* fix collation resolution

* fix #196, fix #195

becauase these have not had FK to the header,
the retention was never applied and now we cannot
simply create the FK as the header does not have all the data.
--truncate table [dbo].[sqlwatch_logger_report_action]
--truncate table [dbo].[sqlwatch_logger_check_action]

* cleanup

* cleanup

* publish profile

* purge orphaned logger records

* added checks template and ability to expand (or explode) check template by an object, for example a single check for database backup can be expanded by each database automatically so we can get descriptive message. This task would be much more suitable for a console app though.

* procedure and query stats

* template actions

* updated grafana dashboards

+++ SqlWatch.Monitor/Project.SqlWatch.Database/dbo/Tables/sqlwatch_config_check_action.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[sqlwatch_config_check_action]
 (
-	[check_id] smallint not null,
+	[check_id] bigint not null,
 	[action_id] smallint not null,
 	[action_every_failure] bit not null constraint df_sqlwatch_config_check_action_every_failure default (0), --whether to send email with every value change as long as its a fail. i.e. we may want to be alerted of every job failure rather than only the first one
 	[action_recovery] bit not null constraint df_sqlwatch_config_check_action_recovery default (1), --whether to send a "recovery" email after check has gone back to OK. just to let us know that there is nothing to worry anymore
